// src/app/pages/covid19/model/formDatosBasicos.model.ts
export class FormDatosBasicos {
    id: number;
    nombre: string;
    apellido: string;
    paisNacionalidad: string;
    ciudadNacimiento: string;
    tipoDocumento: any;
    numeroDocumento: string;
    numeroCelular: string;
    correoElectronico: string;
    ciudadDomicilio: string;
    paisDomicilio: string;
    departamentoDomicilio: string;
    direccionDomicilio: string;
    paisEmisorDocumento: string;
    tipoInicio: string;
    tipoRegistro: string;

    contrasenha: string;
    contrasenhaConfirm: string;

    inicioAislamiento:string;
    fechaPrevistaTomaMuestraLaboratorial:string;
    localTomaMuestra:string;

    //rcToken: string;
    fechaNacimiento:string;

}